! update kode bersama wahyu dan irfan
