+++ shared/src/commonMain/kotlin/com/kompasid/netdatalibrary/core/data/generalContent/network/IGeneralContentApiService.kt
@@ -1,0 +1,9 @@
+package com.kompasid.netdatalibrary.core.data.generalContent.network
+
+import com.kompasid.netdatalibrary.base.network.ApiResults
+import com.kompasid.netdatalibrary.base.network.NetworkError
+import com.kompasid.netdatalibrary.core.data.generalContent.model.dto.GeneralContentResponse
+
+interface IGeneralContentApiService {
+    suspend fun getGeneralContent(): ApiResults<GeneralContentResponse, NetworkError>
+}
