+++ shared/src/commonMain/kotlin/com/kompasid/netdatalibrary/core/data/generalContent/repository/GeneralContentRepository.kt
@@ -4,20 +4,12 @@ import com.kompasid.netdatalibrary.base.network.ApiResults
 import com.kompasid.netdatalibrary.base.network.NetworkError
 import com.kompasid.netdatalibrary.base.network.Results
 import com.kompasid.netdatalibrary.core.data.generalContent.mappers.toInterceptor
-import com.kompasid.netdatalibrary.core.data.generalContent.network.GeneralContentApiService
+import com.kompasid.netdatalibrary.core.data.generalContent.network.IGeneralContentApiService
 import com.kompasid.netdatalibrary.core.domain.generalContent.interceptor.GeneralContentInterceptor
-import com.kompasid.netdatalibrary.core.domain.generalContent.repository.IGeneralContentRepository
 
-// nurirppan__
-// ini nggak pakai interface untuk di unit test?
-// btw gue setuju yu data transform di taruh di repository. namun untuk usecase sendiri kita pakai ini ya
-//✅ Menggabungkan beberapa repository
-//✅ Menjalankan operasi paralel, serial, atau concurrency jika diperlukan
 
-// TODO: Wahyu - Ini interface semua mas, buat yang use case iya nanti bisa di define ulang,
-//  karena general content ini sifatnya standalone gaada bergantung yang lain jadi usecasenya cuma jadi jembatan
 class GeneralContentRepository(
-    private val generalContentServices: GeneralContentApiService
+    private val generalContentServices: IGeneralContentApiService
 ) : IGeneralContentRepository {
 
     override suspend fun getGeneralData(): Results<GeneralContentInterceptor, NetworkError> {
