+++ shared/src/commonMain/kotlin/com/kompasid/netdatalibrary/core/domain/generalContent/usecase/GeneralContentUseCase.kt
@@ -3,7 +3,10 @@ package com.kompasid.netdatalibrary.core.domain.generalContent.usecase
 import com.kompasid.netdatalibrary.base.network.NetworkError
 import com.kompasid.netdatalibrary.base.network.Results
 import com.kompasid.netdatalibrary.core.domain.generalContent.interceptor.GeneralContentInterceptor
+import com.kompasid.netdatalibrary.core.data.generalContent.repository.IGeneralContentRepository
 
-interface GeneralContentUseCase {
-    suspend fun getGeneralData() : Results<GeneralContentInterceptor, NetworkError>
+class GeneralContentUseCase(val repository: IGeneralContentRepository) : IGeneralContentUseCase {
+    override suspend fun getGeneralData(): Results<GeneralContentInterceptor, NetworkError> {
+        return repository.getGeneralData()
+    }
 }
