+++ shared/src/commonMain/kotlin/com/kompasid/netdatalibrary/core/domain/generalContent/usecase/IGeneralContentUseCase.kt
@@ -1,0 +1,9 @@
+package com.kompasid.netdatalibrary.core.domain.generalContent.usecase
+
+import com.kompasid.netdatalibrary.base.network.NetworkError
+import com.kompasid.netdatalibrary.base.network.Results
+import com.kompasid.netdatalibrary.core.domain.generalContent.interceptor.GeneralContentInterceptor
+
+interface IGeneralContentUseCase {
+    suspend fun getGeneralData() : Results<GeneralContentInterceptor, NetworkError>
+}
